// src/app/core/interfaces/helpers/aspx-inputs/admin/block-unblock-user-aspx-input.ts
export interface BlockUnblockUserAspxInput {
  userTypeSelect: HTMLSelectElement;
  userActionSelect: HTMLSelectElement;
  unblockedUsers: HTMLSelectElement;
  blockedUsers: HTMLSelectElement;
  fullNameSelect: HTMLSelectElement;
  reasonForBlock: HTMLInputElement;
  blockSubmitButton: HTMLInputElement;
  unblockSubmitButton: HTMLInputElement;
}